+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Fixed slash commands being shown twice in the menu
 - Some slash commands can now be used in DMs
 - All bot interactions now have cooldowns (3s default)
+- Added `/at-event` which pings all users interested in a given event
 #### HorizonsBot Version 2.1.0:
 - Removed opt-in topics handling, please use the Channel & Roles Browser instead
    - Retained `/petition` for text channels